requirements.txt
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.0
distro==1.9.0
Django==5.1.7
django-celery-beat==2.7.0
django-timezone-field==7.1
dominate==2.9.1
ecdsa==0.19.1
Flask==3.1.0
Flask-APScheduler==1.13.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
kombu==5.4.2
Mako==1.3.9
MarkupSafe==3.0.2
msal==1.31.1
msgspec==0.19.0
oauthlib==3.2.2
openai==1.60.1
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyotp==2.9.0
pyparsing==3.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.2
qrcode==8.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sendgrid==6.12.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
visitor==0.1.3
wcwidth==0.2.13
Werkzeug==3.1.3
